Fully removed ic/. Moved icclient into pfsm/. Removed pfsm/network. Updated imports.
