--- ic/README.md
@@ -1,12 +1,0 @@
-# Internal Communication Packages
-There are 2 packages in ic. icserver and icclient
-
-## icclient
-icclient is used to send messages to the server.
-
-usage can be read in the icclient directory.
-
-## icserver
-icserver is used to listen for messages from clients.
-
-usage can be read in the icserver directory
